move Language model into common.kt
