+++ src/main/kotlin/me/sargunvohra/lib/pokekotlin/model/common.kt
@@ -28,6 +28,15 @@ data class GenerationGameIndex(
     val generation: NamedApiResource
 )
 
+data class Language(
+    val id: Int,
+    val name: String,
+    val official: Boolean,
+    val iso639: String,
+    val iso3166: String,
+    val names: List<Name>
+)
+
 data class Name(
     val name: String,
     val language: NamedApiResource
